requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Excel handling
openpyxl==3.1.2
pandas==2.1.3

# LangChain and AI
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.2
openai==1.6.1

# Web scraping and content extraction
beautifulsoup4==4.12.2
requests==2.31.0
selenium==4.15.2
webdriver-manager==4.0.1

# PDF processing
pypdf2==3.0.1
pdfplumber==0.10.3

# Image processing
pillow==10.1.0
pytesseract==0.3.10

# Environment and configuration
python-dotenv==1.0.0

# Utilities
aiofiles==23.2.1
httpx==0.25.2
